root: backported srcRESEARCH implementations from hestiaERROR
In hestiaERROR repository, there were some improvement works done
for srcRESEARCH/ implementations. Hence, let's backport it.

This patch backports srcRESEARCH implementations from hestiaERROR
into root repository.

+++ automataCI/services/compilers/libreoffice.ps1
@@ -48,14 +48,41 @@ function LIBREOFFICE-Get {
 
 
 
+function LIBREOFFICE-Get-Path-Local {
+	$___path = "${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_TOOLS}\${env:PROJECT_PATH_LIBREOFFICE_ENGINE}\bin\soffice"
+
+	switch ("${env:PROJECT_OS}") {
+	"windows" {
+		$___path = "${___path}.exe"
+	} default {
+		# do nothing
+	}}
+
+
+	# report status
+	return $___path
+}
+
+
+
+
 function LIBREOFFICE-Get-Path {
-	switch ("$(OS-Get)") {
+	# execute
+	$___path = LIBREOFFICE-Get-Path-Local
+	if ($(FS-Is-File "${___path}") -eq 0) {
+		return $___path
+	}
+
+	switch ("${env:PROJECT_OS}") {
 	"darwin" {
 		$___path = "/Applications/LibreOffice.app/Contents/MacOS/soffice"
 	} "windows" {
-		$___path = "C:\Program Files\LibreOffice\program\soffice.exe"
+		$___path = "${___path}.exe"
+		if ($(FS-Is-File "${___path}") -ne 0) {
+			$___path = "C:\Program Files\LibreOffice\program\soffice.exe"
+		}
 	} default {
-		$___path = "${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_TOOLS}\libreoffice\bin\libreoffice"
+		$___path = "/bin/soffice"
 	}}
 
 
@@ -68,8 +95,6 @@ function LIBREOFFICE-Get-Path {
 
 function LIBREOFFICE-Is-Available {
 	# execute
-	$null = OS-Sync
-
 	$___process = OS-Is-Command-Available "libreoffice"
 	if ($___process -eq 0) {
 		return 0
@@ -110,87 +135,67 @@ function LIBREOFFICE-Setup {
 
 
 	# execute
-	if ($(OS-Get) -eq "darwin") {
-		$___dest = "${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_TEMP}\libreoffice\instal.dmg"
-		$null = FS-Make-Housing-Directory "${___dest}"
-		$null = FS-Remove-Silently "${___dest}"
+	$___dest = "${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_TOOLS}\${env:PROJECT_PATH_LIBREOFFICE_ENGINE}"
+	$___url = "${env:PROJECT_LIBREOFFICE_MIRROR}/stable/${env:PROJECT_LIBREOFFICE_VERSION}"
+	switch ("${env:PROJECT_OS}") {
+	"darwin" {
+		$___dest = "${___dest}\install.dmg"
 
-
-		## Download directly from provider
-		$___url = "${env:PROJECT_LIBREOFFICE_MIRROR}/stable/${env:PROJECT_LIBREOFFICE_VERSION}"
 		$___url = "${___url}/mac"
-		if ($(OS-Get-Arch) -eq "amd64") {
+		switch ("${env:PROJECT_ARCH}") {
+		"amd64" {
 			$___url = "${___url}/x86_64"
 			$___url = "${___url}/LibreOffice_${env:PROJECT_LIBREOFFICE_VERSION}_MacOS_x86-64.dmg"
-		} elseif ($(OS-Get-Arch) -eq "arm64") {
+		} "arm64" {
 			$___url = "${___url}/aarch64"
 			$___url = "${___url}/LibreOffice_${env:PROJECT_LIBREOFFICE_VERSION}_MacOS_aarch64.dmg"
-		} else {
-			return 1
+		} default {
+			$___url = ""
+		}}
+
+		## attempt to download from vendor directly
+		if ($(STRINGS-Is-Empty "${___url}") -ne 0) {
+			$null = FS-Make-Housing-Directory "${___dest}"
+			$null = FS-Remove-Silently "${___dest}"
+			$___process = HTTP-Download "GET" "${___url}" "${___dest}"
+			if ($___process -ne 0) {
+				return 1
+			}
+
+			$___process = Apple-Install-DMG "${___dest}"
+			if ($___process -ne 0) {
+				return 1
+			}
+
+			$null = FS-Remove-Silently "${___dest}"
 		}
+	} "windows" {
+		$___dest = "${___dest}\install.msi"
 
-
-		# download from provider
-		$___process = HTTP-Download "GET" "${___url}" "${___dest}"
-		if ($___process -ne 0) {
-			return 1
-		}
-
-
-		## silently install
-		$___process = Apple-Install-DMG "${___dest}"
-		if ($___process -ne 0) {
-			return 1
-		}
-
-
-		## clean up
-		$null = FS-Remove-Silently "${___dest}"
-	} elseif ($(OS-Get) -eq "windows") {
-		## Attempt to use directly from the provider
-		$___dest = "${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_TEMP}\libreoffice\install.msi"
-		$null = FS-Make-Housing-Directory "${___dest}"
-		$null = FS-Remove-Silently "${___dest}"
-
-
-		$___url = "${env:PROJECT_LIBREOFFICE_MIRROR}/stable/${env:PROJECT_LIBREOFFICE_VERSION}"
 		$___url = "${___url}/win"
-		if ($(OS-Get-Arch) -eq "amd64") {
+		switch ("${env:PROJECT_ARCH}") {
+		"amd64" {
 			$___url = "${___url}/x86_64"
 			$___url = "${___url}/LibreOffice_${env:PROJECT_LIBREOFFICE_VERSION}_Win_x86-64.msi"
-		} elseif ($(OS-Get-Arch) -eq "arm64") {
+		} "arm64" {
 			$___url = "${___url}/aarch64"
 			$___url = "${___url}/LibreOffice_${env:PROJECT_LIBREOFFICE_VERSION}_Win_aarch64.msi"
-		} else {
-			## fallback to choco as the last resort
-			$___process = OS-Is-Command-Available "choco"
-			if ($___process -ne 0) {
-				return 1
+		} default {
+			$___url = ""
+		}}
+
+		if ($(STRINGS-Is-Empty "${___url}") -ne 0) {
+			$null = FS-Make-Housing-Directory "${___dest}"
+			$null = FS-Remove-Silently "${___dest}"
+			$___process = HTTP-Download "GET" "${___url}" "${___dest}"
+			if ($___process -eq 0) {
+				$___process = MSI-Install-Silent "${___dest}"
+				$null = FS-Remove-Silently "${___dest}"
 			}
-
-			$___process = OS-Exec "choco" "install libreoffice-fresh -y"
-			if ($___process -ne 0) {
-				return 1
-			}
-
-			return 0
 		}
 
-		$___process = HTTP-Download "GET" "${___url}" "${___dest}"
-		if ($___process -ne 0) {
-			return 1
-		}
-
-		$___process = FS-Is-File "${___dest}"
-		if ($___process -ne 0) {
-			return 1
-		}
-
-		$___process = MSI-Install-Silent "${___dest}"
-		$null = FS-Remove-Silently "${___dest}"
-		$null = OS-Sync
-		if ($___process -ne 0) {
-			## fallback to choco as the last resort
+		## fallback to choco as the last resort
+		if (($(STRINGS-Is-Empty "${___url}") -eq 0) -or ($___process -ne 0)) {
 			$___process = OS-Is-Command-Available "choco"
 			if ($___process -ne 0) {
 				return 1
@@ -202,27 +207,47 @@ function LIBREOFFICE-Setup {
 			}
 
 			$null = OS-Sync
-			return 0
 		}
-	} else {
-		## check compatible platform version
-		$___url = "https://appimages.libreitalia.org"
-		switch ("$(OS-Get-Arch)") {
-		"amd64" {
-			$___url = "${___url}/LibreOffice-fresh.full-x86_64.AppImage"
-		} default {
-			return 1
-		}}
+	} "linux" {
+		$___process = OS-Is-Command-Available "flatpak"
+		if ($___process -eq 0) {
+			$___process = OS-Exec "flatpak" "--user install org.libreoffice.LibreOffice"
+		}
 
-
-		## download appimage portable version
-		$___dest = "${env:PROJECT_PATH_ROOT}\${env:PROJECT_PATH_TOOLS}\libreoffice\bin\libreoffice"
-		$null = FS-Make-Housing-Directory "${___dest}"
-		$null = FS-Remove-Silently "${___dest}"
-		$___process = HTTP-Download "GET" "${___url}" "${___dest}"
+		## download appimage as last resort
 		if ($___process -ne 0) {
-			return 1
+			$___dest = LIBREOFFICE-Get-Path-Local
+			$___url = "https://appimages.libreitalia.org"
+			switch ("${env:PROJECT_ARCH}") {
+			"amd64" {
+				$___url = "${___url}/LibreOffice-fresh.full-x86_64.AppImage"
+			} default {
+				return 1
+			}}
+
+			if ($(STRINGS-Is-Empty "${___url}") -ne 0) {
+				$null = FS-Make-Housing-Directory "${___dest}"
+				$null = FS-Remove-Silently "${___dest}"
+				$___process = HTTP-Download "GET" "${___url}" "${___dest}"
+				if ($___process -ne 0) {
+					return 1
+				}
+
+				$___process = FS-Is-File "${___dest}"
+				if ($___process -ne 0) {
+					return 1
+				}
+			}
 		}
+	} default {
+		return 1 # unsupported
+	}}
+
+
+	# test the output
+	$___process = LIBREOFFICE-Is-Available
+	if ($___process -ne 0) {
+		return 1
 	}
